% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCBM.R
\name{generate_hcbm_correlation_matrix}
\alias{generate_hcbm_correlation_matrix}
\title{Generate Random Correlation Matrix Using Hierarchical Correlation Block Model
You can find the reference for this method here: Reference:  `Marti, G., Andler, S., Nielsen, F. and Donnat, P., 2016.
Clustering financial time series: How long is enough?. arXiv preprint arXiv:1603.04017.
<https://www.ijcai.org/Proceedings/16/Papers/367.pdf>`_}
\usage{
generate_hcbm_correlation_matrix(
  dim,
  block,
  depth,
  lower_bound = 0.1,
  upper_bound = 0.9,
  permute = FALSE
)
}
\arguments{
\item{dim}{an integer indicating the dimension of correlation matrix}

\item{block}{an integer indicating the number of blocks in each recursion}

\item{depth}{an integer showing the number of recursion in hierarchical structure}

\item{lower_bound}{a float in (-1, 1) showing the minimum correlation of assets}

\item{upper_bound}{a float in (0, 1) showing the maximum correlation of assets}

\item{permute}{a boolean to show if we want the order of assets to be randomized}
}
\value{
a random correlation matrix
}
\description{
Generate Random Correlation Matrix Using Hierarchical Correlation Block Model
You can find the reference for this method here: Reference:  `Marti, G., Andler, S., Nielsen, F. and Donnat, P., 2016.
Clustering financial time series: How long is enough?. arXiv preprint arXiv:1603.04017.
<https://www.ijcai.org/Proceedings/16/Papers/367.pdf>`_
}
\examples{
generate_hcbm_correlation_matrix(dim = 50, block = 3, depth = 2)
generate_hcbm_correlation_matrix(dim = 100, block = 2, depth = 3, permute = TRUE)
generate_hcbm_correlation_matrix(dim = 150, block = 3, depth = 3, lower_bound = -0.1)
}
