% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_verification.R
\name{plot_big_eigenvalue_distribution}
\alias{plot_big_eigenvalue_distribution}
\title{Plot the distribution of the k biggest eigenvalues of more than one matrix}
\usage{
plot_big_eigenvalue_distribution(mat, num_of_biggest = NULL)
}
\arguments{
\item{mat}{an dim * dim * n array of n correlation matrices}

\item{num_of_biggest}{an integer showing how many of biggest eigenvalues to be shown}
}
\value{
it plots the distribution of big eigenvalues and doesn't return a value
}
\description{
Plot the distribution of the k biggest eigenvalues of more than one matrix
}
\examples{
plot_big_eigenvalue_distribution(array(rnorm(10 * 10 * 5), dim = c(10, 10, 5)))
plot_big_eigenvalue_distribution(array(rnorm(20 * 20 * 15), dim = c(20, 20, 15)), num_of_biggest = 3)
}
