% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_verification.R
\name{plot_first_eigenvector_entries}
\alias{plot_first_eigenvector_entries}
\title{Plot distribution of the entries of the first eigenvector}
\usage{
plot_first_eigenvector_entries(mat)
}
\arguments{
\item{mat}{a 3d dim * dim * n array of n correlation matrices}
}
\value{
plots the distribution of elements of the first eigenvector and return no value
}
\description{
Plot distribution of the entries of the first eigenvector
}
\examples{
plot_first_eigenvector_entries(array(rnorm(10 * 10 * 5), dim = c(10, 10, 5)))
plot_first_eigenvector_entries(array(rnorm(4 * 4 * 5), dim = c(4, 4, 5)))
}
