% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_verification.R
\name{plot_pairwise_distribution}
\alias{plot_pairwise_distribution}
\title{Plotting the distribution of all pairwise correlations}
\usage{
plot_pairwise_distribution(mat)
}
\arguments{
\item{mat}{a 2d correlation matrix or 3d array of n correlation matrices}
}
\value{
this function just plots. Doesn't return a value
}
\description{
Plotting the distribution of all pairwise correlations
}
\examples{
array(rnorm(10*10), nrow = 10)
plot_pairwise_distribution(array(rnorm(10 * 10 * 5), dim = c(10, 10, 5)))
}
