% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCBM.R
\name{time_series_from_correlation_matrix}
\alias{time_series_from_correlation_matrix}
\title{Generate a Random time series by getting a correlation matrix}
\usage{
time_series_from_correlation_matrix(
  correlation_matrix,
  t_points,
  distribution_type = "normal"
)
}
\arguments{
\item{correlation_matrix}{a correlation matrix. It should be positive semi-definite}

\item{t_points}{an integer showing the number of data points over time to be generated}

\item{distribution_type}{a string indicating the distribution type for asset returns}
}
\value{
a matrix n by t that n is the number of assets and t is the number of datapoints
}
\description{
Generate a Random time series by getting a correlation matrix
}
\examples{
time_series_from_correlation_matrix(mat, 100, 'normal')
time_series_from_correlation_matrix(mat, 50)
time_series_from_correlation_matrix(mat, 150, 'normal')
}
